Extract SubscriptionHandle to separate header
This separate header is leightweight and can be included into ther
headers without a high impact on compilation time. It's mainly intended
for classes with ScopedSubscription members.
